Improve changeset generation strategy
Remove query generator callback and take on the burden of passing the
correct modified and node id clauses.
This may need to be made more flexible in the future, but for now it
fits most use cases.

Remove list of tables. It's now assumed that if changesetQueries is set,
only the tables featured there should be synchronized.

Watched changesets are now generated for the specific modified timestamp
that the change occurred at. This should make those queries more
efficient, and avoids weird edge cases in update storms.

Delta changesets are still generated on connection though.